Camera Control, folder layout and demo scene
I've added a folder layout for us to use that makes everything a bit
easier. Basically, Materials is there for the unity materials (they get
created automatically if you drag a texture onto a gameobject), Models
is there for any 3d models, Prefabs is for any gameobjects that you need
storing (e.g. for instantiation), Scenes is for any scene files (we will
likely just need a menu scene and a game scene), Scripts is for all the
scripts, we may need sub folders if there are a lot of them, and
Textures is for the image files that are used to create a Material (so
.png, .jpg, .psd files)..

I've also created a cameraControl script that is attached to the Camera
Movement gameobject, and thrown together a demo scene to play around
with the camera in (it's hard to test when you can't see depth). The
script is documented if you don't quite understand what's going on.